Bugfixes & improvements
- double instead of float
- Enable greedy heuristics (sorting points)
- Create max_arch() to benchmark against stochastic implementation
- Fix bugs related to comparison operators
- Enable new options
- Set cout precision

+++ head/node.h
@@ -21,13 +21,13 @@
 
 class Node {
 public:
-	float x;
-	float y;
+	double x;
+	double y;
 	Node()
 		: x(0)
 		, y(0)
 	{}
-	Node(float _x, float _y)
+	Node(double _x, double _y)
 		: x(_x)
 		, y(_y)
 	{}
